# crates/manta-cli/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manta API",
    "description": "REST API for managing CSM/OpenCHAMI HPC clusters via manta.",
    "contact": {
      "name": "Manuel Sopena Ballesteros",
      "email": "msopena@cscs.ch"
    },
    "license": {
      "name": ""
    },
    "version": "2.0.0-beta.64"
  },
  "servers": [
    {
      "url": "/api/v2",
      "description": "manta API v2"
    }
  ],
  "paths": {
...
    "/power": {
      "post": {
        "tags": [
          "power"
        ],
        "summary": "`POST /api/v2/power` — start a PCS power transition (on / off /\nreset) against nodes or all members of a cluster and return the\ntransition id **immediately**. Does not block until the\ntransition completes — the CLI is responsible for polling\n`GET /power/transitions/{id}` until the snapshot reports\n`transitionStatus = \"completed\"`.",
        "description": "Returns a `TransitionStartOutput` (`{ transitionID, operation }`)\nas JSON. Callers can hand the `transitionID` to\n[`get_power_transition`].",
        "operationId": "post_power",
        "parameters": [
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PowerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "PCS transition started; returns TransitionStartOutput",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/power/transitions/{id}": {
      "get": {
        "tags": [
          "power"
        ],
        "summary": "`GET /api/v2/power/transitions/{id}` — fetch the current snapshot\nof a PCS power transition (status, task counts, per-task detail).\nCalled by the CLI's poll loop after `POST /power` returns the id.",
        "operationId": "get_power_transition",
        "parameters": [
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transition snapshot",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown transition id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/redfish-endpoints": {
      "get": {
        "tags": [
          "redfish-endpoints"
        ],
        "summary": "GET /redfish-endpoints — list HSM Redfish endpoints with optional filters.",
        "operationId": "get_redfish_endpoints",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Exact endpoint ID (BMC xname) filter.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "fqdn",
            "in": "query",
            "description": "FQDN substring filter.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "uuid",
            "in": "query",
            "description": "UUID exact-match filter.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "macaddr",
            "in": "query",
            "description": "MAC-address exact-match filter (colon-separated hex).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "ipaddress",
            "in": "query",
            "description": "IP-address exact-match filter (IPv4 or IPv6).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Redfish endpoints",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "redfish-endpoints"
        ],
        "summary": "PUT /redfish-endpoints — update an existing Redfish endpoint's properties.",
        "operationId": "update_redfish_endpoint",
        "parameters": [
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRedfishEndpointParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Endpoint updated"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "redfish-endpoints"
        ],
        "summary": "POST /redfish-endpoints — register a new Redfish endpoint in HSM.",
        "operationId": "add_redfish_endpoint",
        "parameters": [
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRedfishEndpointParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Endpoint registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/redfish-endpoints/{id}": {
      "delete": {
        "tags": [
          "redfish-endpoints"
        ],
        "summary": "DELETE /redfish-endpoints/{id} — remove a Redfish endpoint from HSM.",
        "operationId": "delete_redfish_endpoint",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "BMC xname",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Endpoint removed"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatedResponse": {
        "type": "object",
        "description": "Response for endpoints that simply confirm a write happened.\n\nEmitted by `POST /api/v2/redfish-endpoints`,\n`POST /api/v2/boot-parameters`, and `POST /api/v2/groups`.\n\n# Wire shape\n\n```json\n{ \"created\": true }\n```",
        "required": [
          "created"
        ],
        "properties": {
          "created": {
            "type": "boolean",
            "description": "Always `true` on success."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard JSON error body returned by all failed endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable explanation of the failure. Never includes\nstack traces, credentials, or internal type names."
          }
        }
      },
...
      "PowerAction": {
        "type": "string",
        "description": "The power operation to apply to a list of xnames.",
        "enum": [
          "on",
          "off",
          "reset"
        ]
      },
      "PowerRequest": {
        "type": "object",
        "description": "Request body for `POST /api/v2/power`.\n\nThe interpretation of `host_expression` depends on [`PowerTargetType`];\nthe action itself is one of [`PowerAction`]'s variants.\n\n# Wire shape\n\n```json\n{\n  \"action\": \"on\",\n  \"host_expression\": \"x3000c0s1b0n[0-3]\",\n  \"target_type\": \"nodes\",\n  \"force\": false\n}\n```",
        "required": [
          "action",
          "host_expression",
          "target_type"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/PowerAction",
            "description": "Power operation to perform."
          },
          "force": {
            "type": "boolean",
            "description": "Pass `--force` to the underlying power operation (forceful\nshutdown/reset)."
          },
          "host_expression": {
            "type": "string",
            "description": "For `Nodes`: hosts expression (xnames, NIDs, or hostlist\nnotation). For `Cluster`: the HSM group name."
          },
          "target_type": {
            "$ref": "#/components/schemas/PowerTargetType",
            "description": "Whether `host_expression` is a node expression or a cluster\nname."
          }
        }
      },
      "PowerTargetType": {
        "type": "string",
        "description": "Whether the caller's `host_expression` is a hosts expression\n(xnames / NIDs / hostlist) or a single HSM group name whose\nmembers should be targeted.",
        "enum": [
          "nodes",
          "cluster"
        ]
      },
...
      "UpdateRedfishEndpointParams": {
        "type": "object",
        "description": "Typed parameters for adding or updating a Redfish endpoint.\n\nShared request body for **both** `POST /redfish-endpoints` (add a\nnew endpoint) and `PUT /redfish-endpoints` (replace an existing\none by `id`). PUT is a full replacement — every field is written,\nso partial updates require reading the existing endpoint first.",
        "required": [
          "id",
          "enabled",
          "use_ssdp",
          "mac_required",
          "rediscover_on_update"
        ],
        "properties": {
          "domain": {
            "type": [
              "string",
              "null"
            ],
            "description": "Domain portion of the BMC FQDN."
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the endpoint is enabled for discovery."
          },
          "fqdn": {
            "type": [
              "string",
              "null"
            ],
            "description": "Full FQDN; overrides hostname+domain when set."
          },
          "hostname": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hostname portion of the BMC FQDN."
          },
          "id": {
            "type": "string",
            "description": "Xname identifying the BMC (e.g. `x3000c0s1b0`)."
          },
          "ip_address": {
            "type": [
              "string",
              "null"
            ],
            "description": "BMC IP address (IPv4 or IPv6)."
          },
          "mac_addr": {
            "type": [
              "string",
              "null"
            ],
            "description": "BMC MAC address (colon-separated)."
          },
          "mac_required": {
            "type": "boolean",
            "description": "Whether a MAC address is required for geolocation."
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional human-readable name."
          },
          "password": {
            "type": [
              "string",
              "null"
            ],
            "description": "BMC password for Redfish authentication."
          },
          "rediscover_on_update": {
            "type": "boolean",
            "description": "Trigger a rediscovery pass when the endpoint is updated."
          },
          "template_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "ID of a discovery template to apply."
          },
          "use_ssdp": {
            "type": "boolean",
            "description": "Use SSDP for automatic endpoint discovery."
          },
          "user": {
            "type": [
              "string",
              "null"
            ],
            "description": "BMC username for Redfish authentication."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}